Package structure: moved share functionality.
Part of #963.
