+++ src/main/AndroidManifest.xml
@@ -328,7 +328,7 @@ limitations under the License.
         <activity android:name=".settings.SettingsCustomLayoutEditActivity" />
 
         <activity
-            android:name=".ShowOnMapProxyActivity$KML"
+            android:name=".share.ShowOnMapProxyActivity$KML"
             android:exported="true">
             <intent-filter android:label="@string/open_track_as_KML">
                 <action android:name="Intent.OpenTracks-Dashboard" />
@@ -336,7 +336,7 @@ limitations under the License.
             </intent-filter>
         </activity>
         <activity
-            android:name=".ShowOnMapProxyActivity$GPX"
+            android:name=".share.ShowOnMapProxyActivity$GPX"
             android:exported="true">
             <intent-filter android:label="@string/open_track_as_GPX">
                 <action android:name="Intent.OpenTracks-Dashboard" />
