+++ src/main/java/de/dennisguse/opentracks/TrackListActivity.java
@@ -63,6 +63,7 @@ import de.dennisguse.opentracks.services.TrackRecordingServiceConnection;
 import de.dennisguse.opentracks.services.handlers.GpsStatusValue;
 import de.dennisguse.opentracks.settings.PreferencesUtils;
 import de.dennisguse.opentracks.settings.SettingsActivity;
+import de.dennisguse.opentracks.share.ShareUtils;
 import de.dennisguse.opentracks.ui.aggregatedStatistics.AggregatedStatisticsActivity;
 import de.dennisguse.opentracks.ui.aggregatedStatistics.ConfirmDeleteDialogFragment;
 import de.dennisguse.opentracks.ui.markers.MarkerListActivity;
@@ -449,7 +450,7 @@ public class TrackListActivity extends AbstractTrackDeleteActivity implements Co
         }
 
         if (itemId == R.id.list_context_menu_share) {
-            Intent intent = IntentUtils.newShareFileIntent(this, trackIds);
+            Intent intent = ShareUtils.newShareFileIntent(this, trackIds);
             intent = Intent.createChooser(intent, null);
             startActivity(intent);
             return true;
